LayoutTests/fast/forms/watchos/enter-text-with-spellcheck-disabled.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width">
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

async function runTest() {
    if (!window.testRunner) {
        description(`This test requires WebKitTestRunner.`);
        return;
    }

    field = document.getElementById("field");
    await UIHelper.activateElementAndWaitForInputSession(field);
    await UIHelper.enterText("foo bar");
    field.blur();
    await UIHelper.waitForKeyboardToHide();

    shouldBeEqualToString("field.value", "foo bar");
    finishJSTest();
}
</script>
</head>

<body onload="runTest()">
<input id="field" spellcheck="false" style="width: 320px; height: 568px;"></input>
</body>
</html>
